impr: add generic sight
We are also breaking down perspective modules to make it easier to read/navigate.
